Refactor Character Page
